% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Convert back to forward slash}
\usage{
forward(x = NA, writeClipboard = T, return = T)
}
\arguments{
\item{x}{(optional) string - if missing, pulls from clipboard}

\item{writeClipboard}{return string to clipboard?}

\item{return}{return string to R console?}
}
\value{
NULL or string
}
\description{
takes a path from the clipboard or as an argument and pushes to forwardslash
}
\details{
takes a path from the clipboard or as an argument and pushes to forwardslash

\cr
Revision History
\tabular{ll}{
1.2 \tab 2018-03-28  the other \cr
1.1 \tab 2018-03-27  that \cr
1.0 \tab 2018-03-26  this \cr
}
}
\examples{

 writeClipboard("C:\\\\temp\\\\hello")
 forward()
 forward("C:\\\\temp\\\\hello")

}
\seealso{
\code{\link{writeClipboard}}\cr \code{\link{readClipboard}}\cr \code{\link{gsub}}\cr
}
\author{
Jacob Strunk <jstrunk@fs.fed.us>
}
